[syscalls][jobs] Add mx_job_set_relative_importance
Updates a partial ordering between jobs so that job A will be killed before
job B in low-resource situations. Will be used by the out-of-memory (OOM)
ranking service to tell the kernel which jobs to kill first.

No actual OOM killing happens yet.

Also adds the 'k mx jobs' command which dumps the list of all jobs in
importance order.

This reverts commit b92f1a1, which reverted the original commit f97f9fe.
Changes relative to the original commit:
- Avoids wrapping parent_->jobs_ elements in RefPtrs, thus ensures that
  this code won't unintentionally resurrect a job that's in the middle
  of destruction.
- Doesn't hold importance_lock_ as long as before, which isn't critical
  but wasn't necessary.

Change-Id: I82bbf993349004ee8007c28597bc5a995dd15150